requirements.txt
langgraph>=0.2.6
langchain>=0.3.19
langchain-google-genai
langchain-anthropic
langchain-core
pydantic>=2.0.0
python-dotenv>=1.0.1
typing-extensions==4.12.2 